Striot.CompileIoT

Documentation

createPartitions :: StreamGraph -> PartitionMap -> PartitionedGraph

generateCode :: StreamGraph -> PartitionMap -> GenerateOpts -> [String]

Partitions the supplied StreamGraph according to the supplied PartitionMap and options specified within the supplied GenerateOpts and returns a list of the sub-graphs converted into source code and encoded as Strings.

data GenerateOpts

Constructors

GenerateOpts

Fields

imports :: [String]

packages :: [String]

preSource :: Maybe String

rewrite :: Bool

defaultOpts :: GenerateOpts

type PartitionMap = [[Int]]

The user's desired partitioning of the input Graph. Each element in the outer-most list corresponds to a distinct partition. The inner-lists are the IDs of Operators to include in that partition.

writePart :: GenerateOpts -> (Int, String) -> IO ()

genDockerfile :: Bool -> GenerateOpts -> [Char]

partitionGraph :: StreamGraph -> PartitionMap -> GenerateOpts -> IO ()

Partitions the supplied StreamGraph according to the supplied PartitionMap; invokes generateCode for each derived sub-graph; writes out the resulting source code to individual source code files, one per node.

simpleStream :: [(StreamOperator, [ExpQ], String, Double)] -> Graph StreamVertex

Convenience function for specifying a simple path-style of stream processing program, with no merge or join operations. The list of tuples are converted into a series of connected Stream Vertices in a Graph. The tuple arguments are the relevant StreamOperator for the node; the parameters and the *output* type. The other parameters to StreamVertex are inferred from the neighbouring tuples. Unique and ascending vertexId values are assigned.

optimiseWriteOutAll :: FilePath -> [Partition] -> StreamGraph -> IO ()

write out all rewritten versions of the input StreamGraph, along with some of the necessary supporting code. TODO: rename this to something more intuitive

optimise :: StreamGraph -> StreamGraph

Return an optimised version of the supplied StreamGraph, or the graph itself if no better alternative is found.

generateCodeFromStreamGraph :: GenerateOpts -> [(Integer, StreamGraph)] -> StreamGraph -> (Integer, StreamGraph) -> String

nodeFn :: [(Integer, StreamGraph)] -> StreamGraph -> Integer -> StreamGraph -> GenerateOpts -> [Char]

nodeType :: StreamGraph -> NodeType

generateNodeSrc :: Integer -> [Integer] -> GenerateOpts -> [(Integer, StreamGraph)] -> String

connectNodeId :: StreamGraph -> [(Integer, StreamGraph)] -> StreamGraph -> [Integer]

allOptimisations :: StreamGraph -> [(StreamGraph, [OperatorInfo])]

derive all optimisations of the supplied StreamGraph, calculate all Jackson stuff from that

htf_Striot_CompileIoT_thisModulesTests :: TestSuite