Striot.StreamGraph

Documentation

type StreamGraph = Graph StreamVertex

A graph representation of a stream-processing program.

data StreamOperator

An enumeration of the possible stream operators within a stream-processing program, as well as Source and Sink to represent the ingress and egress points of programs.

Constructors

Map

Filter Double

Expand

Window

Merge

Join

Scan

FilterAcc Double

Source Double

Sink

data StreamVertex

The StreamOperator and associated information required to encode a stream-processing program into a Graph. Each distinct StreamVertex within a StreamGraph should have a unique vertexId to ensure that they can be distinguished. For simple path-style graphs, the IDs should be in ascending order.

Constructors

StreamVertex

Fields

vertexId :: Int

operator :: StreamOperator

parameters :: [ExpQ]

intype :: String

outtype :: String

serviceTime :: Double

type PartitionedGraph = ([StreamGraph], StreamGraph)

A collection of partitioned StreamGraphs

deQ :: Q Exp -> Exp

isSource :: StreamOperator -> Bool

showParam :: Q Exp -> String